Changes:
• Renaming Utilities.h to the proper naming convention G_Utilities.h
• Adding in Serialize and Deserilize methods to Utilities.h See TODO: written for more information.
• Adding in recovered IO directory with IO files. (Not implemented).

+++ Project/Common/Window/G_Window.cpp
@@ -62,8 +62,8 @@ namespace Gem
     }
 	void Window::OnWindowResize(float aWidth, float aHeight)
 	{
-		m_Width = aWidth;
-		m_Height = aHeight;
+		m_Width = (int)aWidth;
+		m_Height = (int)aHeight;
 	}
     void Window::Update()
     {
